Return structured data from task
As a plan author I want to consume the result of the `xzscanner::run_scan` task. As such, using structured output for the task will allow me to interact with the data effectively. This commit modifies the task to return structured data such that a plan author can utilize the task more effeciently.